Streamlit/streamlit_app/requirements.txt
pandas
plotly
streamlit
mysql-connector-python>=8.0.32